General code cleanup, removing useless methods, constants and attributes. Updated documentation of some constants inside CloudSimTags in order to provide information about the class of objects for the SimEvent.data attribute when the tag of such an event is set to some of those constantes. Several internal events that were passing an array of IDs now usually pass just one single actual object such as a Host, Vm or Cloudlet. All the additional information that was being passed inside such an array can be got directly from the object passed. For instance, cloudlet.getVm().getHost().getDatacneter() allows to navigate the entire structure from the cloudlet to its datacenter.

+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/datacenters/DatacenterCharacteristics.java
@@ -187,7 +187,7 @@ public interface DatacenterCharacteristics extends Identificable {
      *
      * @return current number of failed PMs the Datacenter has.
      */
-    int getNumberOfFailedHosts();
+    long getNumberOfFailedHosts();
 
     /**
      * Gets the total number of <tt>FREE</tt> or non-busy PEs for all PMs.
@@ -301,7 +301,7 @@ public interface DatacenterCharacteristics extends Identificable {
 
     /**
      * Sets the {@link Datacenter} that owns these characteristics
-     * @param datacenter
+     * @param datacenter the Datacenter to set
      */
     DatacenterCharacteristics setDatacenter(Datacenter datacenter);
 
@@ -342,7 +342,7 @@ public interface DatacenterCharacteristics extends Identificable {
         @Override public int getMips() { return 0; }
         @Override public int getMipsOfOnePe(int hostId, int peId) { return 0; }
         @Override public int getNumberOfBusyPes() { return 0; }
-        @Override public int getNumberOfFailedHosts() { return 0; }
+        @Override public long getNumberOfFailedHosts() { return 0; }
         @Override public int getNumberOfFreePes() { return 0; }
         @Override public int getNumberOfHosts() { return 0; }
         @Override public int getNumberOfPes() { return 0; }
